requirements.txt
anyio==4.9.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2025.7.14
charset-normalizer==3.4.2
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-telegram-bot==13.13
pytz==2025.2
requests==2.32.4
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==1.26.18
